refactor: Update UserCreatedEvent URN and namespace on consumer side message
