+++ services/UserService/src/UserService.Infrastructure/MessageQueue/IdentityService/Constants/EventConstants.cs
@@ -6,7 +6,7 @@ public static class EventConstants
     {
         private const string Prefix = "IdentityService.Exchanges.Users";
         public const string Exchange = $"{Prefix}:UserCreatedExchange";
-        // public const string Urn = $"urn:message:{Exchange}";
+        public const string Urn = $"urn:message:ModCommerce:Events:UserCreated:v1";
 
         private const string QueuePrefix = "UserService.Queues.Users";
         public const string Queue = $"{QueuePrefix}:UserCreatedQueue";
