+++ services/UserService/src/UserService.Infrastructure/MessageQueue/IdentityService/Events/UserCreated/UserCreatedIntegrationEvent.cs
@@ -1,7 +1,8 @@
 using System.Text.Json.Serialization;
 using MassTransit;
 using UserService.Infrastructure.MessageQueue.IdentityService.Constants;
-namespace IdentityService.Contracts.Events.Users;
+
+namespace UserService.Infrastructure.MessageQueue.IdentityService.Events.UserCreated;
 
 /// <summary>
 /// Event published when a new user is created in the Identity service.
@@ -11,7 +12,7 @@ namespace IdentityService.Contracts.Events.Users;
 /// <param name="Username">The username of the created user.</param>
 /// <param name="CreatedAt">The UTC timestamp when the user was created.</param>
 [EntityName(EventConstants.UserCreatedEvent.Exchange)]
-// [MessageUrn(EventConstants.UserCreatedEvent.Urn)]
+[MessageUrn(EventConstants.UserCreatedEvent.Urn)]
 public record UserCreatedEvent(
     /// <summary>
     /// The unique identifier of the created user.
